Move package related models to package module
With the new JSON format we will differentiation between package
vulnerability data and vulnerability advisory data. All package related
data is now moved to the package module. Also add AdvisoryReference
class.
